requirements.txt
gunicorn==21.2.0
asgiref==3.8.1
asttokens==3.0.0
attrs==24.3.0
autobahn==24.4.2
Automat==24.8.1
cffi==1.17.1
channels==3.0.5
colorama==0.4.6
constantly==23.10.4
cryptography==44.0.0
daphne==3.0.2
Django==5.1.4
executing==2.1.0
hyperlink==21.0.0
icecream==2.1.3
idna==3.10
incremental==24.7.2
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
Pygments==2.19.0
pyOpenSSL==24.3.0
service-identity==24.2.0
sqlparse==0.5.3
Twisted==24.11.0
txaio==23.1.1
typing_extensions==4.12.2
tzdata==2024.2
zope.interface==7.2
whitenoise==6.8.2
python-dotenv==1.0.1